Sort out Decrypt()'s API (1st round)
Fix return type; while in there, check for successful decryption.

+++ authenticated_block_ciphers/source/ciphers/AEGIS128_16_128_v1/source/decrypt.c
@@ -125,7 +125,7 @@ static int crypto_aead_decrypt(
 
 
 
-uint8_t Decrypt(uint8_t *block, int32_t mlen, uint8_t *key, uint8_t *npub,
+int Decrypt(uint8_t *block, int32_t mlen, uint8_t *key, uint8_t *npub,
  uint8_t *ad, int32_t  adlen, uint8_t *c, uint8_t *roundKeys)
 {
 	//length of inputs and param
