+++ authenticated_block_ciphers/source/common/cipher.h
@@ -205,7 +205,7 @@ void Encrypt(uint8_t *block, int32_t  mlen, uint8_t *key, uint8_t *npub,
  * ... block - the block to decrypt
  *
  */
-uint8_t Decrypt(uint8_t *block, int32_t  mlen, uint8_t *key, uint8_t *npub,
+int Decrypt(uint8_t *block, int32_t  mlen, uint8_t *key, uint8_t *npub,
  uint8_t *ad, int32_t  adlen, uint8_t *c, uint8_t *roundKeys);
 
 #endif /* CIPHER_H */
